+++ modules/network/basic-cdk/tests/test_stack.py
@@ -1,0 +1,75 @@
+#  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License").
+#    You may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+import os
+
+import aws_cdk as cdk
+import pytest
+from aws_cdk.assertions import Template
+
+
+@pytest.fixture(scope="function")
+def stack_defaults():
+    os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
+    os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
+
+
+def test_synthesize_stack(stack_defaults):
+    import stack
+
+    app = cdk.App()
+    project_name = "test-project"
+    dep_name = "test-deployment"
+    mod_name = "test-module-nointernet"
+    # Create the Stack.
+    network_stack = stack.NetworkingStack(
+        scope=app,
+        id=f"{project_name}-{dep_name}-{mod_name}",
+        project_name=project_name,
+        deployment_name=dep_name,
+        module_name=mod_name,
+        internet_accessible=True,
+        env=cdk.Environment(
+            account=os.environ["CDK_DEFAULT_ACCOUNT"],
+            region=os.environ["CDK_DEFAULT_REGION"],
+        ),
+    )
+    template = Template.from_stack(network_stack)
+    endpoints = template.find_resources(type="AWS::EC2::VPCEndpoint")
+    assert len(endpoints) <= 6
+
+
+def test_synthesize_stack_no_internet(stack_defaults):
+    import stack
+
+    app = cdk.App()
+    project_name = "test-project"
+    dep_name = "test-deployment"
+    mod_name = "test-module-nointernet"
+    # Create the Stack.
+    network_stack = stack.NetworkingStack(
+        scope=app,
+        id=f"{project_name}-{dep_name}-{mod_name}",
+        project_name=project_name,
+        deployment_name=dep_name,
+        module_name=mod_name,
+        internet_accessible=False,
+        env=cdk.Environment(
+            account=os.environ["CDK_DEFAULT_ACCOUNT"],
+            region=os.environ["CDK_DEFAULT_REGION"],
+        ),
+    )
+    template = Template.from_stack(network_stack)
+    endpoints = template.find_resources(type="AWS::EC2::VPCEndpoint")
+    assert len(endpoints) > 6