+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### **Added**
+- adding individual module unit tests, hooked up to workflows
 
 ### **Changed**